Refactor dashboard and remove redundant files
Remove redundant files and code, reorganize files, and update import paths.

* **File Deletions:**
  - Delete `src/app.py`, `src/adware_dashboard/ui/templates/index.html`, and `src/adware_manager.py`.

* **File Renaming and Changes:**
  - Rename `src/adware_dashboard/core/adware_manager.py` to `src/adware_manager.py`.
  - Update import paths in `src/gui.py`, `src/dashboard.py`, `src/adware_dashboard/api/routes.py`, `src/adware_dashboard/api/utils.py`, `src/adware_dashboard/core/ai_integration.py`, `src/adware_dashboard/core/deployment_manager.py`, `src/adware_dashboard/core/payload_manager.py`, and `src/adware_dashboard/ui/views.py` to reflect new locations.

* **Code Reorganization and Enhancements:**
  - Remove redundant imports and code in `src/gui.py`.
  - Integrate `C2Dashboard` class from `src/app.py` into `src/gui.py`.
  - Reorganize functions and methods in `src/gui.py` and `src/dashboard.py` for better readability.
  - Add maximum GUI components to control all functions and configure all possible settings in `src/gui.py`.

* **Error Handling and Validation:**
  - Add error handling for missing API keys in `src/gui.py`.
  - Add validation for empty input data in `src.gui.py`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ProjectZeroDays/zero-click-exploits?shareId=XXXX-XXXX-XXXX-XXXX).
