+++ src/adware_dashboard/api/routes.py
@@ -1,12 +1,12 @@
 import logging
 from flask import Flask, request, jsonify
-from adware_dashboard.core.adware_manager import AdwareManager
-from adware_dashboard.core.payload_manager import PayloadManager
-from adware_dashboard.core.deployment_manager import DeploymentManager
-from adware_dashboard.core.ai_integration import AIIntegration
-from adware_dashboard.models import Adware, Payload, DeploymentMethod
-from adware_dashboard.api.serializers import AdwareSerializer, PayloadSerializer, DeploymentMethodSerializer
-from adware_dashboard.api.utils import validate_input
+from src.adware_manager import AdwareManager
+from src.adware_dashboard.core.payload_manager import PayloadManager
+from src.adware_dashboard.core.deployment_manager import DeploymentManager
+from src.adware_dashboard.core.ai_integration import AIIntegration
+from src.adware_dashboard.models import Adware, Payload, DeploymentMethod
+from src.adware_dashboard.api.serializers import AdwareSerializer, PayloadSerializer, DeploymentMethodSerializer
+from src.adware_dashboard.api.utils import validate_input
 
 def create_api_app(logger: logging.Logger, adware_manager: AdwareManager, payload_manager: PayloadManager, deployment_manager: DeploymentManager, ai_integration: AIIntegration) -> Flask:
     """
