+++ src/adware_dashboard/api/utils.py
