+++ src/adware_dashboard/core/ai_integration.py
