+++ src/adware_dashboard/core/deployment_manager.py
@@ -1,6 +1,6 @@
 import logging
 from typing import List, Dict, Any
-from adware_dashboard.models import DeploymentMethod, Payload
+from src.adware_manager import DeploymentMethod, Payload
 
 class DeploymentManager:
     def __init__(self, logger: logging.Logger):
