+++ src/adware_dashboard/core/payload_manager.py
@@ -1,6 +1,6 @@
 import logging
 from typing import List
-from adware_dashboard.models import Payload
+from src.adware_manager import Payload
 
 class PayloadManager:
     def __init__(self, logger: logging.Logger):
@@ -8,7 +8,7 @@ class PayloadManager:
         Initializes the PayloadManager with a logger.
 
         Args:
-            logger (logging.Logger): The logger instance to use.
+            logger (Logging.Logger): The logger instance to use.
         """
         self.logger = logger
