--- src/adware_dashboard/ui/templates/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Adware Dashboard</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
-</head>
-<body>
-    <h1>Adware Dashboard</h1>
-    <div id="app"></div>
-    <script src="{{ url_for('static', filename='app.js') }}"></script>
-</body>
-</html>
