+++ src/adware_dashboard/ui/views.py
@@ -1,10 +1,10 @@
 from flask import Flask, render_template
-from adware_dashboard.api.routes import create_api_app
-from adware_dashboard.core.adware_manager import AdwareManager
-from adware_dashboard.core.payload_manager import PayloadManager
-from adware_dashboard.core.deployment_manager import DeploymentManager
-from adware_dashboard.core.ai_integration import AIIntegration
-from adware_dashboard.models import create_tables
+from src.adware_dashboard.api.routes import create_api_app
+from src.adware_manager import AdwareManager
+from src.payload_manager import PayloadManager
+from src.deployment_manager import DeploymentManager
+from src.ai_integration import AIIntegration
+from src.models import create_tables
 import logging
 
 logger = logging.getLogger(__name__)
