--- src/app.py
@@ -1,207 +1,0 @@
-import os
-import json
-import logging
-from cryptography.fernet import Fernet
-from ai_model import AIDeploymentModel
-from project_red_sword import Chatbot
-from session_management import SessionManager
-from advanced_decryption import AdvancedDecryption
-from advanced_malware_analysis import AdvancedMalwareAnalysis
-from advanced_social_engineering import AdvancedSocialEngineering
-from adware_manager import AdwareManager
-from ai_red_teaming import AIRedTeaming
-from alerts_notifications import AlertsNotifications
-from android_exploit import AndroidExploit
-from apt_simulation import APTSimulation
-from automated_incident_response import AutomatedIncidentResponse
-from blockchain_logger import BlockchainLogger
-from botnet_manager import BotnetManager
-from data_exfiltration import DataExfiltration
-from data_visualization import DataVisualization
-from deepseek_cody_integration_manager import DeepSeekCodyIntegrationManager
-from device_fingerprinting import DeviceFingerprinting
-from dns_manager import DNSManager
-from download_manager import DownloadManager
-from exploit_payloads import ExploitPayloads
-from fuzzing_engine import FuzzingEngine
-from identity_manager import IdentityManager
-from ios_exploit import IOSExploit
-from iot_exploitation import IoTExploitation
-from linux_exploit import LinuxExploit
-from machine_learning_ai import MachineLearningAI
-from macos_exploit import MacOSExploit
-from mitm_stingray import MITMStingray
-from network_exploitation import NetworkExploitation
-from predictive_analytics import PredictiveAnalytics
-from proxy_chain_manager import ProxyChainManager
-from real_time_monitoring import RealTimeMonitoring
-from real_time_threat_intelligence import RealTimeThreatIntelligence
-from self_healing_ai_manager import SelfHealingAIManager
-from session_management import SessionManagement
-from settings_manager import SettingsManager
-from threat_intelligence import ThreatIntelligence
-from troubleshooting_manager import TroubleshootingManager
-from vscode_dashboard_manager import VSCodeDashboardManager
-from vulnerability_scanner import VulnerabilityScanner
-from windows_exploit import WindowsExploit
-from wireless_exploitation import WirelessExploitation
-from zero_day_exploits import ZeroDayExploits
-
-class C2Dashboard:
-    def __init__(self):
-        self.ai_model = AIDeploymentModel("path/to/pretrained/model.h5")
-        self.chatbot = Chatbot()
-        self.session_manager = SessionManager()
-        self.user_preferences = self.load_user_preferences()
-        self.secure_communication_key = Fernet.generate_key()
-        self.fernet = Fernet(self.secure_communication_key)
-        self.advanced_decryption = AdvancedDecryption()
-        self.advanced_malware_analysis = AdvancedMalwareAnalysis()
-        self.advanced_social_engineering = AdvancedSocialEngineering()
-        self.adware_manager = AdwareManager()
-        self.ai_red_teaming = AIRedTeaming()
-        self.alerts_notifications = AlertsNotifications()
-        self.android_exploit = AndroidExploit()
-        self.apt_simulation = APTSimulation()
-        self.automated_incident_response = AutomatedIncidentResponse()
-        self.blockchain_logger = BlockchainLogger()
-        self.botnet_manager = BotnetManager()
-        self.data_exfiltration = DataExfiltration()
-        self.data_visualization = DataVisualization()
-        self.deepseek_cody_integration_manager = DeepSeekCodyIntegrationManager()
-        self.device_fingerprinting = DeviceFingerprinting()
-        self.dns_manager = DNSManager()
-        self.download_manager = DownloadManager()
-        self.exploit_payloads = ExploitPayloads()
-        self.fuzzing_engine = FuzzingEngine()
-        self.identity_manager = IdentityManager()
-        self.ios_exploit = IOSExploit()
-        self.iot_exploitation = IoTExploitation()
-        self.linux_exploit = LinuxExploit()
-        self.machine_learning_ai = MachineLearningAI()
-        self.macos_exploit = MacOSExploit()
-        self.mitm_stingray = MITMStingray()
-        self.network_exploitation = NetworkExploitation()
-        self.predictive_analytics = PredictiveAnalytics()
-        self.proxy_chain_manager = ProxyChainManager()
-        self.real_time_monitoring = RealTimeMonitoring()
-        self.real_time_threat_intelligence = RealTimeThreatIntelligence()
-        self.self_healing_ai_manager = SelfHealingAIManager()
-        self.session_management = SessionManagement()
-        self.settings_manager = SettingsManager()
-        self.threat_intelligence = ThreatIntelligence()
-        self.troubleshooting_manager = TroubleshootingManager()
-        self.vscode_dashboard_manager = VSCodeDashboardManager()
-        self.vulnerability_scanner = VulnerabilityScanner()
-        self.windows_exploit = WindowsExploit()
-        self.wireless_exploitation = WirelessExploitation()
-        self.zero_day_exploits = ZeroDayExploits()
-
-    def load_user_preferences(self):
-        try:
-            with open('config.json', 'r') as f:
-                return json.load(f)
-        except FileNotFoundError:
-            return {}
-
-    def save_user_preferences(self):
-        with open('config.json', 'w') as f:
-            json.dump(self.user_preferences, f)
-
-    def authenticate_user(self, username, password):
-        return True
-
-    def implement_2fa(self):
-        pass
-
-    def check_session_timeout(self):
-        self.session_manager.check_session_timeout()
-
-    def encrypt_message(self, message):
-        return self.fernet.encrypt(message.encode())
-
-    def decrypt_message(self, encrypted_message):
-        return self.fernet.decrypt(encrypted_message).decode()
-
-    def send_secure_message(self, message):
-        encrypted_message = self.encrypt_message(message)
-        response = requests.post("https://secure-communication.com", data={"message": encrypted_message})
-        return response.status_code
-
-    def render(self):
-        return pn.Column(
-            "### Command and Control Dashboard",
-            pn.pane.Markdown("Welcome to the C2 Dashboard. Here you can manage and monitor your operations."),
-            pn.widgets.Button(name="Start Command", button_type="primary"),
-            pn.widgets.Button(name="Stop Command", button_type="danger"),
-            pn.widgets.DataFrame(name="Command Logs")
-        )
-
-    def predict(self, input_data):
-        return self.ai_model.predict(input_data)
-
-    def scan_targets(self):
-        return self.ai_model.scan_targets()
-
-    def modify_exploits(self, target_info):
-        return self.ai_model.modify_exploits(target_info)
-
-    def deploy_exploit(self, target_info):
-        return self.ai_model.deploy_exploit(target_info)
-
-    def run_post_exploitation_module(self, module_name):
-        pass
-
-    def add_tooltips(self):
-        pass
-
-    def add_help_sections(self):
-        pass
-
-    def add_user_onboarding(self):
-        pass
-
-    def add_in_app_tutorials(self):
-        pass
-
-    def add_feedback_system(self):
-        pass
-
-    def add_animations_transitions(self):
-        pass
-
-    def add_encryption(self):
-        pass
-
-    def integrate_secure_communication(self):
-        pass
-
-    def implement_session_timeout(self):
-        pass
-
-    def add_support_for_more_exploit_types(self):
-        pass
-
-    def integrate_vulnerability_scanner(self):
-        pass
-
-    def implement_reporting_feature(self):
-        pass
-
-    def add_notification_system(self):
-        pass
-
-    def integrate_chatbot_assistant(self):
-        pass
-
-    def add_multimedia_support(self):
-        pass
-
-    def implement_message_encryption(self):
-        pass
-
-    def add_search_feature(self):
-        pass
-
-    def enable_message_reactions(self):
-        pass
