+++ src/automated_incident_response.py
@@ -18,52 +18,42 @@ class AutomatedIncidentResponse:
 
     def handle_malware(self, incident_details):
         logging.info(f"Handling malware incident: {incident_details}")
-        # Placeholder for malware incident response logic
         self.quarantine_system(incident_details["system_id"])
         self.remove_malware(incident_details["system_id"])
 
     def handle_phishing(self, incident_details):
         logging.info(f"Handling phishing incident: {incident_details}")
-        # Placeholder for phishing incident response logic
         self.block_phishing_site(incident_details["url"])
         self.notify_users(incident_details["affected_users"])
 
     def handle_data_breach(self, incident_details):
         logging.info(f"Handling data breach incident: {incident_details}")
-        # Placeholder for data breach incident response logic
         self.secure_system(incident_details["system_id"])
         self.notify_authorities(incident_details["data_type"])
 
     def quarantine_system(self, system_id):
         logging.info(f"Quarantining system: {system_id}")
-        # Placeholder for system quarantine logic
 
     def remove_malware(self, system_id):
         logging.info(f"Removing malware from system: {system_id}")
-        # Placeholder for malware removal logic
 
     def block_phishing_site(self, url):
         logging.info(f"Blocking phishing site: {url}")
-        # Placeholder for phishing site blocking logic
 
     def notify_users(self, affected_users):
         logging.info(f"Notifying affected users: {affected_users}")
-        # Placeholder for user notification logic
 
     def secure_system(self, system_id):
         logging.info(f"Securing system: {system_id}")
-        # Placeholder for system securing logic
 
     def notify_authorities(self, data_type):
         logging.info(f"Notifying authorities about data breach involving: {data_type}")
-        # Placeholder for authority notification logic
 
     def render(self):
         return "Automated Incident Response Module: Ready to respond to and contain security incidents."
 
     def integrate_with_new_components(self, new_component_data):
         logging.info("Integrating with new components")
-        # Placeholder for integration logic with new components
         integrated_data = {
             "new_component_malware_data": new_component_data.get("malware_data", {}),
             "new_component_phishing_data": new_component_data.get("phishing_data", {}),
@@ -73,7 +63,6 @@ class AutomatedIncidentResponse:
 
     def ensure_compatibility(self, existing_data, new_component_data):
         logging.info("Ensuring compatibility with existing incident response logic")
-        # Placeholder for compatibility logic
         compatible_data = {
             "existing_malware_data": existing_data.get("malware_data", {}),
             "existing_phishing_data": existing_data.get("phishing_data", {}),
@@ -85,7 +74,6 @@ class AutomatedIncidentResponse:
         return compatible_data
 
     def update_main_gui(self, incident_type, incident_details):
-        # Placeholder for logic to update the main GUI with incident details
         pass
 
     def integrate_with_main_gui(self, main_gui):
