+++ src/dashboard.py
@@ -44,9 +44,9 @@ from vulnerability_scanner import VulnerabilityScanner
 from windows_exploit import WindowsExploit
 from wireless_exploitation import WirelessExploitation
 from zero_day_exploits import ZeroDayExploits
-from adware_dashboard.core.adware_manager import AdwareManager
-from adware_dashboard.core.ai_integration import AIIntegration
-from adware_dashboard.core.deployment_manager import DeploymentManager
+from adware_manager import AdwareManager
+from ai_integration import AIIntegration
+from deployment_manager import DeploymentManager
 
 class Dashboard:
     def __init__(self, logger: logging.Logger, settings_manager):
