+++ src/data_visualization.py
@@ -50,7 +50,6 @@ class DataVisualization:
         plt.show()
 
     def integrate_with_new_components(self, new_component_data):
-        # Placeholder for integration logic with new components
         integrated_data = {
             "new_component_device_data": new_component_data.get("device_data", {}),
             "new_component_traffic_data": new_component_data.get("traffic_data", {}),
@@ -61,7 +60,6 @@ class DataVisualization:
         return integrated_data
 
     def ensure_compatibility(self, existing_data, new_component_data):
-        # Placeholder for compatibility logic
         compatible_data = {
             "existing_device_data": existing_data.get("device_data", {}),
             "existing_traffic_data": existing_data.get("traffic_data", {}),
